requirements.txt
langchain==0.1.12
chromadb==0.4.24
neo4j==5.17.0
gradio==4.19.2
python-dotenv==1.0.1
pytest==8.0.2
black==24.2.0
flake8==7.0.0
transformers==4.38.2
torch==2.2.1
